combined phil's code with mine
