+++ src/UMLEditor/Style.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package editor;
+package UMLEditor;
 
 public class Style {
     // static fx
