+++ src/UMLEditor/view/Menu.java
@@ -3,8 +3,9 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package editor;
+package UMLEditor.view;
 
+import UMLEditor.Main;
 import javafx.scene.control.SeparatorMenuItem;
 
 /**
